Fixed compiler warnings from VC++.

+++ Development/Source/Core/Geometry/E3GeometryNURBPatch.c
@@ -804,7 +804,7 @@ e3geom_nurbpatch_worldscreen_subdiv( TQ3Point3D** thePoints, TQ3Uns32* numPoints
 	// truncate subdiv into an integer and cache the worldToWindow matrix
 	if( theView ) {
 		// truncate subdiv
-		subdiv = floor( subdiv ) ;
+		subdiv = (float) floor( subdiv ) ;
 		// cache the worldToWindow matrix for theView
 		Q3View_GetWorldToFrustumMatrixState(theView,  &worldToFrustum);
 		Q3View_GetFrustumToWindowMatrixState(theView, &frustumToWindow);
